refactor: Route 경로를 수동으로 업데이트하는 Test Controller 분리
